+++ discord-strava/models/team_leaderboard.rb
@@ -44,17 +44,26 @@ class TeamLeaderboard
     count distance moving_time elapsed_time elevation pr_count calories
   ].freeze
 
-  attr_accessor :team, :metric, :channel_id
+  attr_accessor :team, :metric, :start_date, :end_date, :channel_id
 
   def initialize(team, options = {})
     @team = team
     @metric = options[:metric]
+    @start_date = options[:start_date]
+    @end_date = options[:end_date]
     @channel_id = options[:channel_id]
   end
 
   def aggreate_options
     aggreate_options = { team_id: team.id, _type: 'UserActivity' }
     aggreate_options.merge!('channel_message.channel_id' => channel_id) if channel_id
+    if start_date && end_date
+      aggreate_options.merge!('start_date' => { '$gte' => start_date, '$lte' => end_date })
+    elsif start_date
+      aggreate_options.merge!('start_date' => { '$gte' => start_date })
+    elsif end_date
+      aggreate_options.merge!('start_date' => { '$lte' => end_date })
+    end
     aggreate_options
   end
 
@@ -62,6 +71,7 @@ class TeamLeaderboard
     @aggregate ||= begin
       raise DiscordStrava::Error, "Missing value. Expected one of #{MEASURABLE_VALUES.or}." unless metric && !metric.blank?
       raise DiscordStrava::Error, "Invalid value: #{metric}. Expected one of #{MEASURABLE_VALUES.or}." unless MEASURABLE_VALUES.map(&:downcase).include?(metric.downcase)
+      raise DiscordStrava::Error, 'Invalid date range. End date cannot be before start date.' if @start_date && @end_date && @start_date > @end_date
 
       UserActivity.collection.aggregate(
         [
@@ -105,6 +115,17 @@ class TeamLeaderboard
         rank: row[:rank]
       ).to_s
     }.compact
-    top.any? ? top.join("\n") : "There are no activities with #{metric.split('_').join(' ')} in this channel."
+    if top.any?
+      top.join("\n")
+    else
+      [
+        'There are no activities',
+        metric.downcase == 'count' ? nil : "with #{metric.downcase.split('_').join(' ')}",
+        start_date && end_date ? "between #{start_date.to_fs(:long)} and #{end_date.to_fs(:long)}" : nil,
+        start_date && end_date.nil? ? "after #{start_date.to_fs(:long)}" : nil,
+        start_date.nil? && end_date ? "before #{end_date.to_fs(:long)}" : nil,
+        'in this channel.'
+      ].compact.join(' ')
+    end
   end
 end